small_file_analysis/requirements.txt
awswrangler>=3.4.0
boto3>=1.28.0
pandas>=1.3.0
pyyaml>=6.0
